SoftwareProcess exposes LifecycleEffectorTasks implementation as config
